feat: simplify terminal wrapping component
